feat: built-in endpoints to get sessions, session by id and runs
