+++ libs/agno/agno/os/managers/session/router.py
@@ -3,7 +3,8 @@ from typing import List, Optional, Union
 from fastapi import APIRouter, HTTPException, Path, Query
 
 from agno.db.base import BaseDb, SessionType
-from agno.os.managers.session.schemas import (
+from agno.os.managers.utils import PaginatedResponse, PaginationInfo, SortOrder
+from agno.os.schema import (
     AgentSessionDetailSchema,
     DeleteSessionRequest,
     RunSchema,
@@ -13,7 +14,6 @@ from agno.os.managers.session.schemas import (
     WorkflowRunSchema,
     WorkflowSessionDetailSchema,
 )
-from agno.os.managers.utils import PaginatedResponse, PaginationInfo, SortOrder
 
 
 def attach_routes(router: APIRouter, db: BaseDb) -> APIRouter:
