+++ libs/agno/agno/os/managers/session/session.py
@@ -1,6 +1,5 @@
 import logging
 from typing import Optional
-from uuid import uuid4
 
 from fastapi.routing import APIRouter
